+++ src/components/data/converters/address.js
@@ -2,22 +2,20 @@ import _ from 'lodash';
 import { hashCode } from 'hashcode';
 
 function toListAddress(addr) {
-  if (addr.id) return { ...addr };
-
   const country = (addr.countryId) ? addr.countryId : '';
-  const id = hashCode().value(addr).toString();
+  const addressId = addr.id || hashCode().value(addr).toString();
 
   return {
-    id,
+    id: addressId,
     addressLine1: addr.addressLine1,
     addressLine2: addr.addressLine2,
     city: addr.city,
-    primaryAddress: addr.primaryAddress,
-    primary: addr.primaryAddress,
     stateRegion: addr.region,
     zipCode: addr.postalCode,
     country,
     addressType: addr.addressTypeId,
+    primaryAddress: addr.primaryAddress,
+    primary: addr.primaryAddress,
   };
 }