+++ src/components/data/converters/address.test.js
@@ -9,7 +9,7 @@ describe('getFormAddressList', () => {
     expect(getFormAddressList([])).toEqual([]);
   });
 
-  test('maps object fields correctly when they lack "id" property', () => {
+  test('should map object fields correctly when they lack "id" property', () => {
     const expected = [
       {
         addressLine1: 'a.al1',
@@ -63,18 +63,30 @@ describe('getFormAddressList', () => {
   });
 
 
-  test('returns objects as-is if they contain "id" property', () => {
+  test('should map object fields correctly when they contain "id" property', () => {
     const expected = [
       {
         addressLine1: 'a.al1',
         addressLine2: 'a.al2',
-        postalCode: 'a.postalCode',
+        city: 'a.city',
+        primaryAddress: 2,
+        primary: 2,
+        stateRegion: 'a.region',
+        zipCode: 'a.postalCode',
+        addressType: 'a.addressTypeId',
+        country: 'a.countryId',
         id: 'a.id',
       },
       {
         addressLine1: 'b.al1',
         addressLine2: 'b.al2',
-        postalCode: 'b.postalCode',
+        city: 'b.city',
+        primaryAddress: 1,
+        primary: 1,
+        stateRegion: 'b.region',
+        zipCode: 'b.postalCode',
+        addressType: 'b.addressTypeId',
+        country: '',
         id: 'b.id',
       }
     ];
@@ -82,13 +94,22 @@ describe('getFormAddressList', () => {
       {
         addressLine1: 'a.al1',
         addressLine2: 'a.al2',
+        city: 'a.city',
+        primaryAddress: 2,
+        region: 'a.region',
         postalCode: 'a.postalCode',
+        addressTypeId: 'a.addressTypeId',
+        countryId: 'a.countryId',
         id: 'a.id',
       },
       {
         addressLine1: 'b.al1',
         addressLine2: 'b.al2',
+        city: 'b.city',
+        primaryAddress: 1,
+        region: 'b.region',
         postalCode: 'b.postalCode',
+        addressTypeId: 'b.addressTypeId',
         id: 'b.id',
       },
     ];